extractors/corp_portal/bsc_standards.py
# extractors/corp_portal/bsc_standards.py
import os
from datetime import datetime
from dotenv import load_dotenv

import sys
from pathlib import Path

load_dotenv(dotenv_path=".env")

def run():
    download_dir = os.getenv("DOWNLOAD_DIR", "data/raw/bsc_standards")
    os.makedirs(download_dir, exist_ok=True)

    print(f"[INFO] BSC Standards extractor - placeholder implementation")
    print(f"[INFO] Would download to: {download_dir}")

    # TODO: Implement full extraction logic
    return True

if __name__ == "__main__":
    success = run()
    sys.exit(0 if success else 1)