requirements.txt
# ===========================================================================
# UNIFIED NORTHLIGHT PLATFORM DEPENDENCIES
# Consolidated from Heartbeat and Northlight projects
# ===========================================================================

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==22.0.0
starlette==0.47.2

# Database Drivers and ORMs
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Data Processing (consolidated versions)
pandas==2.3.1
numpy==2.3.2
pyarrow==17.0.0

# Legacy Database Support (for migration)
duckdb==1.0.0

# Web Scraping and Automation
playwright==1.47.0
selenium==4.24.0
beautifulsoup4==4.12.3
lxml==6.0.0
html5lib==1.1

# Office Document Processing
openpyxl==3.1.5
xlsxwriter==3.2.5
python-pptx==1.0.2
python-docx==1.1.2

# Authentication and Security
authlib==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# API and HTTP
httpx==0.27.0
aiohttp==3.10.5
requests==2.32.3
httptools==0.6.4

# Validation and Serialization
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.14.1

# Configuration and Environment
python-dotenv==1.1.1
pydantic-settings==2.5.2
dynaconf==3.2.6

# Date and Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Caching and Session Management
redis==5.0.8
aioredis==2.0.1

# Monitoring and Logging
structlog==24.4.0
rich==13.8.1
loguru==0.7.2

# Task Queue and Background Jobs
celery[redis]==5.4.0
kombu==5.4.2

# Image Processing
pillow==11.3.0

# Data Validation and Quality
cerberus==1.3.5
jsonschema==4.23.0

# Utilities
click==8.2.1
colorama==0.4.6
six==1.17.0
idna==3.10
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0
watchfiles==1.1.0
websockets==15.0.1

# Development and Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2

# File Format Support
PyYAML==6.0.2
toml==0.10.2
et-xmlfile==2.0.0

# Encryption and Hashing
cryptography==43.0.1
bcrypt==4.2.0

# Async Support
asyncio-mqtt==0.16.2
aiofiles==24.1.0

# Data Export and Reporting
jinja2==3.1.4
matplotlib==3.9.2
seaborn==0.13.2

# API Documentation
openapi-spec-validator==0.7.1

# Performance Monitoring
prometheus-client==0.20.0

# Notebook Support (for data analysis)
jupyter==1.1.1
ipython==8.27.0

# Machine Learning (if needed for analytics)
scikit-learn==1.5.2
scipy==1.14.1

# TOTP/2FA Support
pyotp==2.9.0
qrcode==8.0

# Excel Advanced Features
xlwings==0.32.2

# Email Support
aiosmtplib==3.0.2
email-validator==2.2.0

# Type Inspection
typing-inspection==0.4.1